I restructured the decision tree within find_ball_groups so that it looks at color first, then moves on to distance. This seems to work much better. It seems to be able to group balls moving very quickly, even if there is a missing frame. Huge improvement.
I also added a method to help with motion-blurred circles. Often Hough-
Circles detects 2 circles on a ball when it is moving very quickly, but
so far the circles have always somewhat overlapped, so I added a method
that would look at all the circles from the frame, and see if any were
closer together than their radii would allow in real life, and if there
were, I would replace them with a point right in between the two. This
improves color comparison as well.

I also made a few of the Table methods static, and rearranged them to be
all together at the bottom of the Table methods.