Coolest thing: I increased the fps from ~20 to ~35 by having the table's find_balls method crop down the playfield image, rather than assigning zeros. I had to add the cropped amount back into the circle centers after, but it's way, way faster. Very happy about that.
I have also started adding the trajectory estimation lines. Currently,
it takes the balls past centers and fits a line to them, then calculates
where the hit-point would be on the wall. Have not yet factored in the
other balls or reflection angles. That's coming. I added a movement
threshold attribute to the balls that determines how far the ball has to
go in a frame to draw the motion line. Also added an ismoving attribute
which I think I will need when it comes to collision estimation. The
ball trajectory estimation is currently all happening in the Ball's
find_future_path method. I also had cProfile sort by time.

Issues:
- false balls still appear on hands and clothes, and they often appear
randomly, triggering the drawing of trajectory lines. That's the main
issue at the moment.